+++ libs/dimcli/cli.cpp
@@ -2402,25 +2402,31 @@ static bool parse(Cli & cli, vector<string> & args) {
 }
 
 //===========================================================================
-bool Cli::parse(size_t argc, char * argv[]) {
-    auto args = toArgv(argc, argv);
+bool Cli::parse(vector<string> & args) {
     return ::parse(*this, args);
 }
 
 //===========================================================================
 bool Cli::parse(vector<string> && args) {
-    return ::parse(*this, args);
+    return parse(args);
+}
+
+//===========================================================================
+bool Cli::parse(size_t argc, char * argv[]) {
+    auto args = toArgv(argc, argv);
+    return parse(move(args));
+}
+
+//===========================================================================
+bool Cli::parse(size_t argc, const char * argv[]) {
+    auto args = toArgv(argc, argv);
+    return parse(move(args));
 }
 
 //===========================================================================
 bool Cli::parse(const vector<string> & args) {
     auto tmp = args;
-    return ::parse(*this, tmp);
-}
-
-//===========================================================================
-bool Cli::parse(vector<string> & args) {
-    return ::parse(*this, args);
+    return parse(move(tmp));
 }
 
 //===========================================================================