+++ CHANGELOG.md
@@ -11,8 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## Unreleased
-- Added - cli.success() and cli.fail() overload with just the exit code
-- Added - cli.parse() overload taking a const string vector
+- Added - cli.success() and cli.fail() overloads with just the exit code
+- Added - cli.parse() const overloads for vector and argv
 
 ## dimcli 7.4.0 (2025-06-16)
 - Added - cli.beforeExec() and cli.afterExec() for code common to all commands